Update ml-template.cpp
Update t type from int to long long.

+++ Cf_cpp_template/ml-template.cpp
@@ -9,7 +9,7 @@ int main() {
 #endif
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
-    int t;
+    ll t;
     cin >> t;
     while (t--) {
